+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -853,9 +853,9 @@ abstract class ImageProxy {
 /// Utilities for working with [ImageProxy]s.
 @ProxyApi()
 abstract class ImageProxyUtils {
-  /// Returns a single buffer that is representative of three NV21-compatible [planes].
+  /// Returns a single buffer that is representative of three NV21-compatible planes.
   @static
-  Uint8List getNv21Buffer(int imageWidth, int imageHeight, List<PlaneProxy> planes);
+  Uint8List getNv21Buffer(ImageProxy imageProxy);
 }
 
 /// A plane proxy which has an analogous interface as